requirements.txt
Django>=5.1.2,<6.0
djangorestframework>=3.15.2,<4.0
psycopg2>=2.9,<3.0
drf-spectacular>=0.27.2,<0.29
djangorestframework-simplejwt>=5.3.1,<5.9
Pillow>=11.0.0,<12.0.0